Refactorización y nuevas funcionalidades en Inquilinos, Contratos y Pagos.
Gestión de Contratos:

Se implementa la lógica (Contratos.java, ContratosFragment.java, ContratosViewModel.java, DetalleContratosFragment.java, DetalleContratosViewModel.java) y la interfaz de usuario (fragment_detalle_contratos.xml, contrato_card.xml) para la administración de contratos.

Gestión de Inquilinos:

Se añade el módulo completo para inquilinos (Inquilino.java, InquilinosFragment.java, InquilinosViewModel.java, DetalleInquilinosFragment.java, DetalleInquilinosViewModel.java) incluyendo la UI (fragment_detalle_inquilinos.xml).

Gestión de Pagos:

Se integra el módulo de pagos (Pagos.java, PagosFragment.java, PagosViewModel.java) con sus respectivos adaptadores y vistas (PagosAdapter.java, pagos_card.xml, fragment_pagos.xml).

Gestión de Inmuebles:

Se incluye lógica para la carga de inmuebles (CargarInmueblesFragment.java, CargarInmueblesViewModel.java) y la entidad principal (Inmueble.java).
Arquitectura: Se sigue un patrón de diseño ViewModel (MVVM) en todos los módulos, como se evidencia en los múltiples archivos ...ViewModel.java.

Adaptadores (UI/Datos): Se crearon adaptadores dedicados (ContratosAdapter.java, PagosAdapter.java, InquilinosAdapter.java) para manejar las listas y la visualización de datos en la interfaz.

Conectividad: Se actualiza el cliente de API (ApiClient.java) para soportar las nuevas entidades y operaciones.

Estructura Base: Se actualizan archivos clave del core de la aplicación: LoginActivity.java, LogoutFragment.java, MainActivity.java, app_bar_main.xml, y la navegación principal (mobile_navigation.xml).

+++ Inmobiliaria JD/app/src/main/java/com/example/inmobiliariajd/model/Inmueble.java
@@ -8,7 +8,7 @@ public class Inmueble implements Serializable {
     private String uso;
     private String tipo;
     private int ambientes;
-    private double superficie;
+    private int superficie;
     private double latitud;
     private double longitud;
     private double valor;
@@ -17,7 +17,7 @@ public class Inmueble implements Serializable {
     private int idPropietario;
     private Propietario duenio;
 
-    public Inmueble(int idInmueble, String direccion, String uso, String tipo, int ambientes, double superficie,
+    public Inmueble(int idInmueble, String direccion, String uso, String tipo, int ambientes, int superficie,
                     double latitud, double longitud, double valor, String imagen, boolean disponible,
                     int idPropietario, Propietario duenio) {
         this.idInmueble = idInmueble;
@@ -79,11 +79,11 @@ public class Inmueble implements Serializable {
         this.ambientes = ambientes;
     }
 
-    public double getSuperficie() {
+    public int getSuperficie() {
         return superficie;
     }
 
-    public void setSuperficie(double superficie) {
+    public void setSuperficie(int superficie) {
         this.superficie = superficie;
     }
